requirements.txt
aiogram==3.15.0
python-dotenv==1.1.0
aiohttp==3.11.14
